Add files via upload
This project is an Amazon clone created using HTML, CSS, and JavaScript. The aim was to replicate the basic structure and functionality of Amazon's online shopping platform. Key features include a responsive design, product listings, a shopping cart, and user authentication. By utilizing HTML for the structure, CSS for styling, and JavaScript for interactivity, this clone provides a comprehensive look at how a major e-commerce site operates. This endeavor not only enhances web development skills but also offers valuable insights into the complexities of building a functional and visually appealing online store.